change how normalizr detects pagination
Problem: Our change detection failed if response
contained no items i.e. empty array

Solution: Only change if the totalPages is defined
and a value greater or equal to 0. This catches
all the cases where totalPages isnt defined,
because there is no pagination, and also
when there is pagination, but there are 0 or more
pages.