change style of follow up answer form
Problem: there is a lot of code here when it can mostly
be handled by redux form

Solution: Create a more standardised handling of these
answer properties by always looking at it from a question
level. An element is now passed into the forms and then we
select a specific question out of that to edit in the form.