delete questionnaireFolders from cache when deleting
Problem: When we delete a folder,
we dont delete the questionnaireFolders which stores
the mapping from a questionnaire, to a folder. This
means that in the UI, it sees the questionnaire
as mapped to a folder that no longer exists.

Solution: We can clear out all of the questionnaireFolders
from the cache that are for this folder after we
do the folder deletion.

close #312